Push some depedency and properties to FileUtility and create failing test suite for extension checking
