+++ client/src/Service/File/MimeTypeAndExtensionChecker.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Service\File;
+
+class MimeTypeAndExtensionChecker extends FileUtility
+{
+    public function check(): bool
+    {
+        return true;
+    }
+}
